# Use in-memory atom ordering for map ordering