This should optimize all operations on small maps
that rely on ordering. As an example, creating a
map with 32 keys named 'atom-INDEX', where INDEX
is an integer, is an order of magnitude faster
with this patch.

Before:

    from_list      624.32 K        1.60 μs   ±723.44%        1.54 μs        1.71 μs

After:

    from_list        6.23 M      160.47 ns ±15898.23%         125 ns         167 ns

It has been verified that the distribution and
ETS will respect the atom ordering of the system
and therefore are backwards compatible.

A potential downside is that the order of atom
fields in maps is no longer deterministic, but
that's already the case for large maps, and having
non-deterministic atom ordering may help avoid
code that accidentally relies on order.

This patch also opens up the possibility for
more efficient map lookups in the future.