@josevalim 2022-07-13
I am pretty sure CI will break, but I am sending a PR to assess how big the damage will be. :)