@vans163 2022-07-13
Could this be a emulator runtime or compile time flag? I can think of a few cases where the order of atoms is very much appreciated, mostly in tooling, such as not having to manually override print functions of maps/structs (members automatically sorted) when inspecting them in logs/console.

Also would this affect non-atom keyed maps or mix of key type maps?

```
and having non-deterministic atom ordering may help avoid code that accidentally relies on order.
```
Well it wont help if you know your never expecting a map larger than a certain size. Say your parsing protobuf->map you know the packet will only ever have 4 fields based on the protodefs, and reading the packetlog with the fields always in different places wont work for debugging, so youl have to write a custom print function. Again its kind of a time saver for an untyped language, as it should be, poc and iterate asap.