@josevalim 2022-07-13
> Also would this affect non-atom keyed maps or mix of key type maps?

No, it should not.

> Well it wont help if you know your never expecting a map larger than a certain size. 

Future Erlang/OTP versions can reduce the size or even change the small map implementation altogether to one that does not rely on ordering. So I can understand the benefit of ordering for debugging/printing, but for anything else you should not rely on it.

For those cases, however, maybe we should consider looking for solutions that work both for small and large maps alike. For example, maybe we could have a flag on `io:format/2` that orders the keys before-hand.