@vans163 2022-07-13
> Future Erlang/OTP versions can reduce the size or even change the small map implementation altogether to one that does not rely on ordering. So I can understand the benefit of ordering for debugging/printing, but for anything else you should not rely on it.

Indeed, its more of a convenience.

> For printing, however, maybe we should consider looking for solutions that work both for small and large maps alike. For example, maybe we could have a flag on `io:format/2` that orders the keys before-hand.

Yea that would work well, then the `defimpl Inspect` can be overwritten with said `io:format` flag so the REPL print shows it ordered too (on Elixir side, not sure how this is done on Erlang side).

The speed difference is very significant.