@bjorng 2022-07-27
Thanks for your pull request.

I think that the OTP Technical Board will have to decide whether this change is acceptable from a backward-compatibility perspective. There could be applications or libraries that depend on `maps:to_list(SmallMap)` always returning a sorted list. I have no idea common that could be.

Regarding printing, I think that the easiest way to handle it is to modify `io:format()` to always sort maps when printing them. I am not sure whether we need an option to print maps in the raw order. The OTB will have to decide on how to handle printing, too.

There is one bug introduced by this PR, namely that `term_to_binary(SmallMapWithAtomKeys, [deterministic])` is broken. That bug will also make BEAM files non-deterministic because the compiler uses `term_to_binary/2` with the `deterministic` option.

Your PR got me thinking about alternative ways to speed up comparisons of atoms. More about my ideas in the next comment.