@bjorng 2022-07-27
Here is the code that compares atoms:

https://github.com/erlang/otp/blob/93f8daa84b351e1ba4f135bec1b7803004dd3a34/erts/emulator/beam/erl_utils.h#L188-L215

If the first 4 characters (more correct: 3 characters and 7 bits) in the atoms to be compared are different, the comparison can be done relatively cheap without calling `memcmp()`. But if they are equal, `memcmp()` need to be called.

If the keys are all atoms named `'atom-INDEX'` as in your commit message, `memcmp()` will always have to be called. One idea for optimisations would be store the 8 first characters (strictly speaking: 7 characters and 7 bits) for each atom in the atom entry. For many atoms named `'atom-INDEX'`, `memcmp()` would not have to be called. The cost for that would be 4 more bytes for each atom. (On a 32-bit machine, I suggest keeping the current implementation.)

I have not done any measurements to find out how much faster that would be compared to the current implementation.