@josevalim 2022-07-27
Thank you @bjorng!

---

Comments related to the board discussion:

> There could be applications or libraries that depend on maps:to_list(SmallMap) always returning a sorted list. I have no idea common that could be.

To play devil's advocate: we may have other reasons to change how SmallMaps are implemented in the future, so the sooner we make it unordered, theoretically the better. Although I understand even then it may be better to just wait if the need arises (it might never!).

> Regarding printing, I think that the easiest way to handle it is to modify io:format() to always sort maps when printing them.

My concern with sorting on printing is when you convert it to list or iterate it, you will see a different order. :(

---

Comments about optimizations:

> If the keys are all atoms named 'atom-INDEX' as in your commit message, memcmp() will always have to be called. One idea for optimisations would be store the 8 first characters (strictly speaking: 7 characters and 7 bits) for each atom in the atom entry.

Oh, interesting. So I accidentally triggered a bad path!

I also wonder if we could further optimize ASCII atoms? Use 1 bit to denote if ASCII or not. If ASCII, then you can pack the double amount of characters? The downside is that comparing an ASCII atom with a non-ASCII atom will always trigger the slow path but I am not sure how common that is.