@bjorng 2022-07-27
> I also wonder if we could further optimize ASCII atoms? Use 1 bit to denote if ASCII or not. If ASCII, then you can pack the double amount of characters?

No, you can't gain that much. Atoms are stored in UTF8, which means that ASCII characters (code points 0 through 127) are already stored in only one byte. You could do some tricks such as only storing 7 bits from each ASCII character and fit 9 characters into 63 bits, or even more by subtracting off 32 from each character and multiplying by 96 instead of left-shifting 7 positions. I doubt that the more complicated code is worth it.

(I will take two days off and be back on Monday.)