@bjorng 2022-08-15
@josevalim Could you modify your benchmark to only use unique atoms as maps keys?

That could give us some idea of how expensive the call to `strcmp()` is and whether we should optimise map comparisons in the way I suggested.